Extract default AggregateRoot implementation into its own trait.
